feat(ga): update list change
