+++ packages/webapp/src/webview/ui/components/App/App.scss
@@ -109,7 +109,7 @@ h1 {
     margin-top: 20px;
 }
 
-.striped-list-bookmarks {
+.striped-list-items {
     padding-left: 0;
     margin-top: 20px;
     > li:nth-of-type(odd) {
